functions/gcf_training_classification/requirements.txt
# GCP sdk
google-cloud-storage==1.42.1
google-cloud-aiplatform==1.1.1